refactor(rtp): extract the RtpSession inbound pipeline into RtpInboundProcessor (#14)
RtpSession sat at 999 lines — one under the hard 1000-line file limit, so any
further growth would break the architecture gate. Move the inbound half — the
RFC 7983/5764 demux, SRTP/SRTCP authenticate-and-decrypt, RFC 3550 §A.1 sequence
validation, fail-closed drops and dispatch (ProcessDatagram + the secondary RTX
path) — into a dedicated internal RtpInboundProcessor.

The session keeps the socket, the send path, the symmetric-RTP latch and the
§8.2 SSRC-collision reseed (send-side sequence/timestamp/SSRC ownership); the
processor owns the inbound security contexts and the secondary-stream routing
state and dispatches back through the session's events via ctor-injected
callbacks, so PacketReceived keeps the session as its sender and teardown's
event-clearing is honoured. A detected collision routes to the session via a
callback that runs synchronously on the same receive-loop thread — no new race.

Pure extract-class: the moved body is line-for-line identical, the external
surface (all six events + InstallSecurityContexts/InstallSecondarySecurityContexts/
ConfigureSecondaryStream/SecondaryPayloadType/InjectInboundDatagramForTest) is
unchanged, the Volatile inbound-context discipline is preserved, and the logger
category stays RtpSession. RtpSession 999→731, processor 393 — both under limit.

Release 0 warn · Arch 12/12 (net10) · Core 1557/1557 (net9) · Client 89/89.
Revert-verify: breaking the PacketReceived dispatch wiring turned the primary-
delivery tests RED, restoring turned them GREEN — the extracted path is live.
